ASB-267 Rename examination to projects, rename about-the-application to documents
